levels/level_07_multi_agent/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-system"
version = "0.1.0"
description = "Agentic RAG System - local-only, OSS"
requires-python = ">=3.12"
dependencies = [
    # Vector DB
    "weaviate-client>=4.9",
    # Embeddings
    "sentence-transformers>=3.3",
    # LLM Client (LM Studio compatible)
    "openai>=1.50",
    # Agent Graph
    "langgraph>=0.2",
    # Chunking
    "langchain-text-splitters>=0.3",
    "beautifulsoup4>=4.12",
    # PDF / PPTX
    "pymupdf>=1.24",
    "python-pptx>=1.0",
    # Config / Data Models
    "pyyaml>=6.0",
    "pydantic>=2.9",
    # UI
    "reflex>=0.6",
    # ColBERT (optional, may need Python 3.12 workaround)
    # "ragatouille>=0.0.9",
    # Re-ranker (included in sentence-transformers)
    # Testing
    "pytest>=8.0",
    "pytest-mock>=3.14",
    "requests>=2.31",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
